Feature/authentication

+++ src/app/users/users.controller.ts
@@ -1,8 +1,27 @@
-import { BadRequestException, Body, Controller, Delete, Get, Param, Post, Put } from '@nestjs/common';
+import {
+    BadRequestException,
+    Body,
+    ClassSerializerInterceptor,
+    Controller,
+    Delete,
+    ForbiddenException,
+    Get,
+    NotFoundException,
+    Param,
+    Post,
+    Put,
+    Req,
+    UseGuards,
+    UseInterceptors,
+} from '@nestjs/common';
+import { FastifyRequest } from 'fastify';
+import { AuthenticationGuard } from '../authentication';
 import { DmaLogger } from '../logging';
-import { CreateUserData, User } from './models';
+import { CreateUserData, UpdateUserData, User } from './models';
 import { UsersService } from './users.service';
 
+@UseGuards(AuthenticationGuard)
+@UseInterceptors(ClassSerializerInterceptor)
 @Controller('users')
 export class UsersController {
     constructor(
@@ -20,18 +39,26 @@ export class UsersController {
 
     @Post()
     public async create(@Body() userData: CreateUserData) {
-        this.logger.log('Creating a new user', userData);
+        this.logger.log('Creating a new user');
         return await this.usersService.create(userData);
     }
 
-    @Get(':id')
+    @Get('/:id')
     public async getById(@Param('id') userIdParam: string) {
         this.logger.log(`Getting user with ID "${userIdParam}"`);
-        return await this.usersService.getById(userIdParam);
+        const query = await this.usersService.getById(userIdParam);
+
+        if (!query) throw new NotFoundException(`User with ID "${userIdParam}" was not found`);
+        return query;
     }
 
-    @Put(':id')
-    public async update(@Param('id') userIdParam: string, @Body() userData: User) {
+    @Put('/:id')
+    public async update(
+        @Param('id') userIdParam: string,
+        @Body() userData: UpdateUserData,
+        @Req() request: FastifyRequest
+    ) {
+        this.validateResourceOwner(userIdParam, request.raw.authenticatedUser);
         this.logger.log(`Update user with ID "${userIdParam}"`);
 
         if (userIdParam !== userData.id) {
@@ -42,9 +69,17 @@ export class UsersController {
         return await this.usersService.update(userData);
     }
 
-    @Delete(':id')
-    public async removeByid(@Param('id') userIdParam: string) {
+    @Delete('/:id')
+    public async removeById(@Param('id') userIdParam: string, @Req() request: FastifyRequest) {
+        this.validateResourceOwner(userIdParam, request.raw.authenticatedUser);
         this.logger.log(`Remove user with ID "${userIdParam}"`);
+
         await this.usersService.removeById(userIdParam);
+    }
+
+    private validateResourceOwner(userId: string, authenticatedUser: User) {
+        if (userId === authenticatedUser.id) return;
+        this.logger.warn(`Unauthorized access attempt detected by User with ID "${authenticatedUser.id}"`);
+        throw new ForbiddenException('You are not authorized to perform this action');
     }
 }
